refactor: replace `Reflection` with `type erased interfaces`
Wanted it to see if this design would simplify things. I still don't
know how I feel about it, but I think I like it???

refactor: `OnFinishedSession()` sets `UserSettings` to `default`

I still think this is a bad idea, but for now I want to keep it like
this and see if this causes any problems.

refactor: improve logs when loading and saving settings